pyproject.toml
[tool.poetry]
name = "blogapp"
version = "0.1.0"
description = ""
authors = ["abhishek.vaish <abhishek.vaish@acsicorp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
email-validator = "^1.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.23.0"
requests = "^2.28.1"
pytest-asyncio = "^0.19.0"
asgi-lifespan = "^1.0.1"


[tool.poetry.group.dev.dependencies]
tox = "^3.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile="black"


[tool.black]
multi_line_output = 3
include_trailing_comma= true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88


[tool.bandit]
exclude_dirs = [
    "./.eggs",
    "./.git",
    "./.tox",
    "./build",
    "./coverage",
    "./dist",
    "./venv",
    "./__pycache__",
    "*tests*",
]
